more work to nwe interface
