+++ js/ogsip.js
@@ -62,6 +62,7 @@ function hashchange() {
 	last_letter = letter;
     _gaq.push(['_trackPageview', '/' + window.location.hash.substring(1)]);
 }
+
 function loadLetter(letter, game_name) {
 	$("#games").load("modules/games.php?letter=" + letter, function(game) {
         var radio = document.getElementById("radio"+letter);
